Generate colors in a Tailwind.Theme module.
